feat: add Cytoscape context and provider for graph management
