+++ backhaul-xcien/package.json
@@ -25,6 +25,7 @@
     "@react-aria/visually-hidden": "3.8.22",
     "@react-types/shared": "3.29.0",
     "clsx": "2.1.1",
+    "cytoscape": "^3.32.0",
     "framer-motion": "11.15.0",
     "react": "18.3.1",
     "react-dom": "18.3.1",
