+++ backhaul-xcien/src/main.tsx
@@ -1,17 +1,20 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
 import { BrowserRouter } from "react-router-dom";
-
 import App from "./App.tsx";
-import { Provider } from "./provider.tsx";
+import { HeroUIProviderWrapper, CytoscapeProvider } from "@/providers";
+
 import "@/styles/globals.css";
+
 
 ReactDOM.createRoot(document.getElementById("root")!).render(
   <React.StrictMode>
     <BrowserRouter>
-      <Provider>
-        <App />
-      </Provider>
+      <HeroUIProviderWrapper>
+        <CytoscapeProvider>
+          <App />
+        </CytoscapeProvider>
+      </HeroUIProviderWrapper>
     </BrowserRouter>
   </React.StrictMode>,
 );
