+++ backhaul-xcien/src/providers/HeroUIProviderWrapper.tsx
@@ -1,0 +1,22 @@
+import type { NavigateOptions } from "react-router-dom";
+
+import { HeroUIProvider } from "@heroui/system";
+import { useHref, useNavigate } from "react-router-dom";
+
+declare module "@react-types/shared" {
+  interface RouterConfig {
+    routerOptions: NavigateOptions;
+  }
+}
+
+export function HeroUIProviderWrapper({ children }: { children: React.ReactNode }) {
+  const navigate = useNavigate();
+
+  return (
+    <HeroUIProvider navigate={navigate} useHref={useHref}>
+      {/* <main className="dark text-foreground bg-background"> */}
+      {children}
+      {/* </main> */}
+    </HeroUIProvider>
+  );
+}
